Module rupta::util

Modules

bit_vec

Our implemention of bit vector is modifed from rustc_index::bit_set, see https://doc.rust-lang.org/stable/nightly-rustc/src/rustc_index/bit_set.rs.html

call_graph_stat

chunked_queue

dot

Simple graphviz dot file format output. Adapted from dot.rs in petgraph.

index_tree

mem_watcher

Memory usage monitoring. Currently only supported on Linux.

options

Analysis options.

pta_statistics

results_dumper

type_util

unsafe_statistics

Functions

customize_generic_args

find_sysroot

Returns the location of the rust system binaries that are associated with this build of rust-pta. The location is obtained by looking at the contents of the environmental variables that were set at the time rust-pta was compiled. If the rust compiler was installed by rustup, the variables RUSTUP_HOME and RUSTUP_TOOLCHAIN are used and these are set by the compiler itself. If the rust compiler was compiled and installed in some other way, for example from a source enlistment, then the RUST_SYSROOT variable must be set in the environment from which rust-pta is compiled.

has_self_parameter

Returns true if the function has an explicit self (either self or &(mut) self) as its first parameter, allowing method calls.

has_self_ref_parameter

Returns true if the function has an explicit &(mut) self as its first parameter, allowing method calls.

is_dynamic_call

Returns true if the call to (callee_def_id, callee_substs) is a dynamic call.

is_std_lib_func

Returns true if the function identified by def_id is defined in the Rust Standard Library.

is_trait_method

Returns true if the function identified by def_id is defined as part of a trait.

pretty_print_mir

Dumps a human readable MIR redendering of the function with the given DefId to standard output.

qualified_path_to_offset_path

Returns an offset_path equivalent to the qualified_path.