Module rupta::util::type_util

Structs

FieldByteOffsetCache

Provides a way to effectively get the byte offsets of an ADT type’s fields

PathCastCache

Manage the type cast for paths

PointerProjectionsCache

Provides a way to effectively get the pointer type fields of a given type

TypeCache

Provides a way to refer to a rustc_middle::ty::Ty via a handle that does not have a life time specifier.

Functions

are_concrete

Returns false if any of the generic arguments are themselves generic

closure_field_types

Closures bring enclosed variables with them that are effectively additional parameters. There is no convenient way to look up their types later on. I.e. unlike ordinary parameters whose types can be looked up in mir.local_decls, these extra parameters need their types extracted from the closure type definitions via the tricky logic below.

closure_return_type

equal_types

Returns true if the two given types are equal after erasing regions

equivalent_ptr_types

Returns true if the given two pointer types are equivalent. We suppose that a reference type and a mut/const raw pointer type are equivalent if their dereference types are equivalent.
Pointers of equivalent types can point to the same object.

eval_local_decl_type

field_ty

fields_at_start_location

flatten_fields

Given an object that may contain nested objects, flatten it by extracting all the bottom-level subobjects. This function returns a vector of tuples, each including a subobject’s memory offset from the base object, its path representation and its type.

function_return_type

get_array_length

get_dereferenced_type

Returns the target type of a reference type.

get_downcast_type

Returns the rustc TyKind of the downcast projection

get_element_type

Returns the element type of an array or slice type.

get_field_type

Returns the type of the field with the given ordinal.

get_pointer_projections

Returns all the projections of pointer type fields contained in the given type

is_argumentv1_array

is_basic_pointer

Returns whether the type is a pointer to a basic type e.g. *u8, *u8, *()

is_basic_type

Returns whether the type is a primitive type or an array or slice containing basic ty elements e.g. u8, u8, ()

is_concrete

Determines if the given type is fully concrete.

is_dynamic_fn_trait

Returns true if this type is dyn Fn, dyn FnMut or dyn FnOnce.

is_dynamic_pointer

Returns true if the given type is a reference (or raw pointer) to a dynamic type

is_fn_once_call_once

is_fn_once_output

is_fn_trait

Returns true if this id corresponds to the fn_trait|fn_mut_trait|fn_once_trait

is_foreign_pointer

Returns true if the given type is a reference (or raw pointer) to a foreign type

is_slice_pointer

Returns true if the given type is a reference (or raw pointer) to a collection type, in which case the reference/pointer independently tracks the length of the collection, thus effectively tracking a slice of the underlying collection.

is_str_ref_array

is_transparent_wrapper

layout_of

Returns a layout for the given type, if concrete.

matched_fn_sig

projections_and_types

Returns a vector of field projections paired with their corresponding types contained in the given type

remove_transparent_wrapper

repr(transparent) is used on structs with a single non-zero-sized field (there may be additional zero-sized fields). Get the type and field index after removing the transparent wrapper

size_of

Returns the size for the given type

strip_auto_traits

try_eval_path_type