Struct rupta::builder::fpag_builder::FuncPAGBuilder

pub struct FuncPAGBuilder<'pta, 'tcx, 'compilation> { /* private fields */ }

Implementations

impl<'pta, 'tcx, 'compilation> FuncPAGBuilder<'pta, 'tcx, 'compilation>

A visitor that traverses the MIR associated with a particular function’s body and build the corresponding function pag.

pub fn new( acx: &'pta mut AnalysisContext<'tcx, 'compilation>, func_id: FuncId, mir: &'tcx Body<'tcx>, fpag: &'pta mut FuncPAG ) -> FuncPAGBuilder<'pta, 'tcx, 'compilation>

pub fn build(&mut self)

pub fn visit_body(&mut self)

pub fn visit_const_operand(&mut self, const_op: &ConstOperand<'tcx>) -> Rc<Path>

Returns a value that corresponds to the given literal

pub fn visit_const(&mut self, c: &Const<'tcx>) -> Rc<Path>

Synthesizes a constant value from a RustC constant as used in the type system

pub fn visit_unevaluated_const( &mut self, unevaluated: &UnevaluatedConst<'tcx>, ty: Ty<'tcx> ) -> Rc<Path>

Synthesizes a constant value from an unevaluated mir constant which is not part of the type system.

pub fn inline_indirectly_called_function( &mut self, callee_def_id: &DefId, gen_args: &GenericArgsRef<'tcx>, args: Vec<Rc<Path>>, destination: Rc<Path>, location: Location )

Fn::call, FnMut::call_mut, FnOnce::call_once all receive two arguments:

Operand of any type that implements Fn|FnMut|FnOnce, a function pointer or closure instance for most cases.

A tuple of argument values for the call. The tuple is unpacked and the callee is then invoked with its normal function signature. In the case of calling a closure, the closure signature includes the closure as the first argument.

All of this happens in code that is not encoded as MIR, so we need built in support for it.

pub fn add_internal_edges( &mut self, src_path: Rc<Path>, src_type: Ty<'tcx>, dst_path: Rc<Path>, dst_type: Ty<'tcx> )

If the source path and destination path are both of pointer types, add a direct edge between them. Otherwise, get the pointer fields if exist and add internal edges between the pointer fields.

pub fn create_aux_local(&mut self, ty: Ty<'tcx>) -> Rc<Path>

Creates an auxiliary local variable with the given type. Returns the path of the auxiliary local variable.

pub fn get_param_env(&self) -> ParamEnv<'tcx>

Returns a parameter environment for the current function.

pub fn copy_and_transmute( &mut self, source_path: Rc<Path>, source_rustc_type: Ty<'tcx>, target_path: Rc<Path>, target_rustc_type: Ty<'tcx> )

Copy the value at source_path to a value at target_path. If the type of value at source_path is different from that at target_path, the value is transmuted.

pub fn get_box_pointer_field( &mut self, box_path: Rc<Path>, ty: Ty<'tcx> ) -> Rc<Path>

Return the path of the pointer at field 0 of the NonNull pointer at field 0 of the Unique pointer at field 0 of the box create an auxiliary variable to represent this sub-field.

pub fn add_addr_edge(&mut self, src: Rc<Path>, dst: Rc<Path>)

pub fn add_direct_edge(&mut self, src: Rc<Path>, dst: Rc<Path>)

pub fn add_store_edge(&mut self, src: Rc<Path>, dst: Rc<Path>)

For (*p).f1.f2…fn = q, add a Store edge ‘q –STORE(f1.f2…fn)–> p’

pub fn add_load_edge(&mut self, src: Rc<Path>, dst: Rc<Path>)

For p = (*q).f1.f2…fn, add a Load edge ‘q –LOAD(f1.f2…fn)–> p’

pub fn add_gep_edge(&mut self, src: Rc<Path>, dst: Rc<Path>)

For p = &((*q).f1.f2…fn), add a Gep edge ‘q –GEP(f1.f2…fn)–> p’

pub fn add_cast_edge(&mut self, src: Rc<Path>, dst: Rc<Path>)

pub fn add_offset_edge(&mut self, src: Rc<Path>, dst: Rc<Path>)

pub fn add_edge(&mut self, src: Rc<Path>, dst: Rc<Path>, kind: PAGEdgeEnum)

Adds an internal edge from src to dst of kind to the function pag.