Type Alias rupta::graph::call_graph::CSCallGraph

pub type CSCallGraph = CallGraph<CSFuncId, CSBaseCallSite>;

Aliased Type

struct CSCallGraph {
    pub graph: Graph<CallGraphNode<CSFuncId>, CallGraphEdge<BaseCallSiteS<CSFuncId>>>,
    pub func_nodes: HashMap<CSFuncId, NodeIndex>,
    pub callsite_to_edges: HashMap<BaseCallSiteS<CSFuncId>, HashSet<EdgeIndex>>,
    /* private fields */
}

Fields
graph: Graph<CallGraphNode<CSFuncId>, CallGraphEdge<BaseCallSiteS<CSFuncId>>>

The graph structure capturing calls between nodes

func_nodes: HashMap<CSFuncId, NodeIndex>

A map from function value to call graph node id

callsite_to_edges: HashMap<BaseCallSiteS<CSFuncId>, HashSet<EdgeIndex>>

A map from call sites to call graph edges