Struct rupta::graph::pag::PAG

pub struct PAG<P: PAGPath> { /* private fields */ }

Implementations

impl<P: PAGPath> PAG<P>

pub fn new() -> Self

Constructor

pub fn graph(&self) -> &Graph<PAGNode<P>, PAGEdge>

Returns a reference to the pag graph.

pub fn addr_edge_iter(&self) -> IterCopied<PAGEdgeId>

Return a iterator for the address_of edges.

pub fn node_path(&self, node_id: PAGNodeId) -> &P

Returns the path for the given node_id

pub fn get_node(&self, node_id: PAGNodeId) -> &PAGNode<P>

Returns the node for the given node_id

pub fn get_node_mut(&mut self, node_id: PAGNodeId) -> &mut PAGNode<P>

Returns the node for the given node_id

pub fn get_node_id(&self, path: &P) -> Option<PAGNodeId>

Returns the node_id for the given path

pub fn get_edge(&self, edge_id: PAGEdgeId) -> &PAGEdge

Returns the edge for the given edge_id

pub fn add_node(&mut self, path: P)

Adds a new node to the pag.

pub fn get_or_insert_node(&mut self, path: &P) -> PAGNodeId

Helper function to get a node or insert a new node if it does not exist in the map.

pub fn has_edge(&self, src: &P, dst: &P, kind: &PAGEdgeEnum) -> bool

Returns true if the edge from src path to dst path of the kind exits.

pub fn contains_edge( &self, src: PAGNodeId, dst: PAGNodeId, kind: &PAGEdgeEnum ) -> bool

Returns true if the edge from src to dst of the kind exits.

pub fn add_incoming_addr_edge(&mut self, node_id: PAGNodeId, in_edge: PAGEdgeId)

pub fn add_outgoing_addr_edge( &mut self, node_id: PAGNodeId, out_edge: PAGEdgeId )

pub fn add_incoming_direct_edge( &mut self, node_id: PAGNodeId, in_edge: PAGEdgeId )

pub fn add_outgoing_direct_edge( &mut self, node_id: PAGNodeId, out_edge: PAGEdgeId )

pub fn add_incoming_load_edge(&mut self, node_id: PAGNodeId, in_edge: PAGEdgeId)

pub fn add_outgoing_load_edge( &mut self, node_id: PAGNodeId, out_edge: PAGEdgeId )

pub fn add_incoming_store_edge( &mut self, node_id: PAGNodeId, in_edge: PAGEdgeId )

pub fn add_outgoing_store_edge( &mut self, node_id: PAGNodeId, out_edge: PAGEdgeId )

pub fn add_incoming_gep_edge(&mut self, node_id: PAGNodeId, in_edge: PAGEdgeId)

pub fn add_outgoing_gep_edge(&mut self, node_id: PAGNodeId, out_edge: PAGEdgeId)

pub fn add_incoming_cast_edge(&mut self, node_id: PAGNodeId, in_edge: PAGEdgeId)

pub fn add_outgoing_cast_edge( &mut self, node_id: PAGNodeId, out_edge: PAGEdgeId )

pub fn add_incoming_offset_edge( &mut self, node_id: PAGNodeId, in_edge: PAGEdgeId )

pub fn add_outgoing_offset_edge( &mut self, node_id: PAGNodeId, out_edge: PAGEdgeId )

pub fn add_edge( &mut self, src: &P, dst: &P, kind: PAGEdgeEnum ) -> Option<PAGEdgeId>

Adds an edge from src to dst according to the edge type. Returns the edge id if this edge is newly added to the graph

pub fn add_addr_edge(&mut self, src: &P, dst: &P) -> Option<PAGEdgeId>

pub fn add_direct_edge(&mut self, src: &P, dst: &P) -> Option<PAGEdgeId>

pub fn add_load_edge( &mut self, src: &P, dst: &P, kind: PAGEdgeEnum ) -> Option<PAGEdgeId>

pub fn add_store_edge( &mut self, src: &P, dst: &P, kind: PAGEdgeEnum ) -> Option<PAGEdgeId>

pub fn add_gep_edge( &mut self, src: &P, dst: &P, kind: PAGEdgeEnum ) -> Option<PAGEdgeId>

pub fn add_cast_edge(&mut self, src: &P, dst: &P) -> Option<PAGEdgeId>

pub fn add_offset_edge(&mut self, src: &P, dst: &P) -> Option<PAGEdgeId>

pub fn add_new_direct_edges<'tcx>( &mut self, acx: &mut AnalysisContext<'tcx, '_>, src: &P, dst: &P, ty: Ty<'tcx> ) -> Vec<PAGEdgeId>

Given two paths, add direct edge between them if they are of pointer type or add direct edges between their pointer type fields if any. Return the edges added.

pub fn get_func_pag(&self, func_id: &FuncId) -> Option<&FuncPAG>

pub fn build_func_pag( &mut self, acx: &mut AnalysisContext<'_, '_>, func_id: FuncId ) -> bool

pub fn build_static_pag( &mut self, acx: &mut AnalysisContext<'_, '_>, def_id: DefId ) -> Option<FuncId>

pub fn promote_constants<'tcx>( &mut self, acx: &mut AnalysisContext<'tcx, '_>, def_id: DefId, gen_args: &Vec<GenericArgE<'tcx>> ) -> Option<HashSet<FuncId>>

impl<P: PAGPath> PAG<P>
where P::FuncTy: Into<FuncId> + Copy,

pub fn add_inter_procedural_edges( &mut self, acx: &mut AnalysisContext<'_, '_>, callsite: &CallSiteS<P::FuncTy, P>, callee: P::FuncTy ) -> Vec<PAGEdgeId>

Adds direct edges from the arguments to the parameters and from the return value to the destination value.