Trait rupta::mir::path::PathSupport

pub trait PathSupport {
    // Required methods
    fn is_field_of(&self, path: &Rc<Path>) -> bool;
    fn is_deref_path(&self) -> bool;
}

cannot define inherent impl for a type outside of the crate where the type is defined define and implement a trait or new type instead

Required Methods

fn is_field_of(&self, path: &Rc<Path>) -> bool

fn is_deref_path(&self) -> bool